requirements.txt
watchdog>=3.0.0,<4.0.0
pytest>=7.0.0
toml>=0.10.2
pyyaml>=6.0
questionary>=2.0.0
ruamel.yaml>=0.17
